bmv2 switched to 32-bit table entry handles

+++ p4c_bm/templates/src/pd_ageing.cpp
@@ -81,7 +81,7 @@ void ${pd_prefix}ageing_notification_cb(const char *hdr, const char *data) {
             << "for table " << ageing_hdr->table_id << std::endl;
     const AgeingState *state = device_state[ageing_hdr->switch_id];
     const AgeingCb &cb = state->cbs.find(ageing_hdr->table_id)->second;
-    const uint64_t *handles = reinterpret_cast<const uint64_t *>(data);
+    const uint32_t *handles = reinterpret_cast<const uint32_t *>(data);
     for(unsigned int i = 0; i < ageing_hdr->num_entries; i++) {
       cb.cb_fn(handles[i], cb.cb_cookie);
     }
